requirements.txt
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.1
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cache2==0.2.2
greenlet==3.2.1
h11==0.14.0
httptools==0.6.4
idna==3.10
kombu==5.5.3
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
